# [shellcraft] Avoid recursive walk of all templates for command line