This mechanism is much faster (up to ~5x)

Before: `pwn shellcraft i386.linux.sh  0.57s user 1.47s system 20% cpu 10.060 total`
After: `pwn shellcraft i386.linux.sh  0.42s user 0.27s system 25% cpu 2.722 total`

This also has the  side-effect of allowing e.g. `shellcraft amd64.open` which mimics the in-Python behavior:

```
$ shellcraft amd64.open '/dev/null' 0 0
6a6c48b82f6465762f6e756c504889e731d231f66a02580f05
```

Fixes #1650