@heapcrash 2020-07-28
Lint failed, but it failed because we *removed* a warning.  I need to add logic to that checker to only show added warnings.